+++ src/initialize/Radius.js
@@ -1,28 +1,48 @@
-import Initialize from './Initialize';
+import Initializer from './Initializer';
 import { createSpan } from '../math';
 
-export default class Radius extends Initialize {
+/**
+ * Sets the radius property on initialized particles.
+ *
+ */
+export default class Radius extends Initializer {
   /**
-   * Radius is init particle's Radius
-   * @param {Number} a - the Radius's start point
-   * @param {Number} b - the Radius's end point
-   * @param {String} c - span's center
-   * @example
-   * var Radius = new Proton.Radius(3,5);
-   * or
-   * var Radius = new Proton.Radius(3,1,"center");
-   * @extends {Initialize}
-   * @constructor
+   * Constructs a Radius initializer instance.
+   *
+   * @param {number} width - The width of the particle radius
+   * @param {number} height - The height of the particle radius
+   * @param {boolean} [center=false] - Determines whether to average the radius value
+   * @return void
    */
-  constructor(a, b, c) {
+  constructor(width, height, center = false) {
     super();
-    this.radius = createSpan(a, b, c);
+
+    /**
+     * @desc The radius span which is used to set the particle radius value.
+     * @type {Span}
+     */
+    this.radius = createSpan(width, height, center);
   }
 
-  reset(a, b, c) {
-    this.radius = createSpan(a, b, c);
+  /**
+   * Resets the initializer properties.
+   * Clears all previously set zones and resets the zones according to args passed.
+   *
+   * @param {number} width - The width of the particle radius
+   * @param {number} height - The height of the particle radius
+   * @param {boolean} [center=false] - Determines whether to average the radius value
+   * @return void
+   */
+  reset(width, height, center = false) {
+    this.radius = createSpan(width, height, center);
   }
 
+  /**
+   * Sets the particle's initial radius.
+   *
+   * @param {Particle} particle - the particle to initialize the property on
+   * @return void
+   */
   initialize(particle) {
     particle.radius = this.radius.getValue();
     particle.transform.oldRadius = particle.radius;