+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.8] - 2018-12-02
+
+### Added
+
+- Test coverage for all particle initializers
+
+### Changed
+
+- The base `Initialize` class is now `Initializer`, will change the `initialize` module to `initializer` when some re-architecting is done
+
 ## [1.0.7] - 2018-12-01
 
 ## Added